Optimise analyse function
Before, the analyse function would calculate the number of literals in the current decision level by looping through the learned clause over every iteration. This has been changed, so now whenever a literal from the current decision level is added, the counter is increased, and whenever a literal from the current decision level is removed, the counter is decreased.